logistics_api/microservices/tracking_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
pydantic
python-dotenv
pydantic-settings
email-validator